+++ src/testing.ts
@@ -3,3 +3,17 @@
  */
 export { callContract, contractHandlerMetaKey } from './testing/callContract'
 export type { ContractHandlerMeta } from './testing/callContract'
+export { testContract, ContractAssertionError } from './testing/contractTest'
+export type {
+  ContractTestSuite,
+  ContractTestInput,
+  TestHandler,
+} from './testing/contractTest'
+export {
+  startContractCoverage,
+  stopContractCoverage,
+  resetContractCoverage,
+  getContractCoverage,
+  formatContractCoverage,
+} from './testing/coverage'
+export type { ContractCoverageReport, ContractCoverageEntry } from './testing/coverage'